EE-9765 Code review change: moved IncomeSummary into new application.domain package
