+++ src/main/java/uk/gov/digital/ho/pttg/api/HmrcResource.java
@@ -4,7 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.bind.annotation.*;
 import uk.gov.digital.ho.pttg.application.NinoUtils;
-import uk.gov.digital.ho.pttg.dto.IncomeSummary;
+import uk.gov.digital.ho.pttg.application.domain.IncomeSummary;
 import uk.gov.digital.ho.pttg.dto.Individual;
 
 import java.time.Instant;
