+++ src/main/java/uk/gov/digital/ho/pttg/application/domain/IncomeSummary.java
@@ -1,8 +1,12 @@
-package uk.gov.digital.ho.pttg.dto;
+package uk.gov.digital.ho.pttg.application.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import uk.gov.digital.ho.pttg.dto.AnnualSelfAssessmentTaxReturn;
+import uk.gov.digital.ho.pttg.dto.Employment;
+import uk.gov.digital.ho.pttg.dto.Income;
+import uk.gov.digital.ho.pttg.dto.Individual;
 
 import java.util.List;
