+++ src/test/java/uk/gov/digital/ho/pttg/api/HmrcResourceIntegrationTest.java
@@ -22,7 +22,7 @@ import org.springframework.web.client.RestTemplate;
 import uk.gov.digital.ho.pttg.ServiceRunner;
 import uk.gov.digital.ho.pttg.application.HmrcAccessCodeClient;
 import uk.gov.digital.ho.pttg.dto.AccessCode;
-import uk.gov.digital.ho.pttg.dto.IncomeSummary;
+import uk.gov.digital.ho.pttg.application.domain.IncomeSummary;
 
 import java.io.IOException;
 import java.math.BigDecimal;
